requirements.txt
python-telegram-bot
SQLAlchemy
Pillow
fastapi
uvicorn